Add userId to return of Login and Register routes. Login and Register front-end now use context to set token and userId. Add dummy logout button to Home.js for testing
